requirements.txt
awsebcli==3.4.6
backports.ssl-match-hostname==3.4.0.2
boto==2.38.0
cement==2.4.0
cffi==1.1.2
cryptography==0.9.1
cssselect==0.9.1
Django==1.8.2
django-endless-pagination==2.0
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
enum34==1.0.4
idna==2.0
ipaddress==1.0.7
jmespath==0.7.1
lxml==3.4.4
pathspec==0.3.3
Pillow==2.8.2
psycopg2==2.6
pyasn1==0.1.7
pycparser==2.14
pyOpenSSL==0.15.1
python-dateutil==2.4.2
PyYAML==3.11
queuelib==1.2.2
requests==2.6.2
Scrapy==0.24.6
six==1.9.0
SQLAlchemy==1.0.5
texttable==0.8.3
Twisted==15.2.1
w3lib==1.11.0
websocket-client==0.32.0
wheel==0.24.0
zope.interface==4.1.2
slugify==0.0.1
django-tinymce==1.5.3
twython==3.2.0
